完善手动更新的exmple

+++ Utils/library/src/main/java/com/library/bean/Version.java
@@ -2,7 +2,7 @@ package com.library.bean;
 
 public class Version {
 
-	private Integer versionID;
+	private Float versionID;
 	private String downloadURL;
 	private String describe;
 	public String getDescribe() {
@@ -19,11 +19,11 @@ public class Version {
 		this.downloadURL = downloadURL;
 	}
 
-	public Integer getVersionID() {
+	public Float getVersionID() {
 		return versionID;
 	}
 
-	public void setVersionID(Integer versionID) {
+	public void setVersionID(Float versionID) {
 		this.versionID = versionID;
 	}
 }
